+++ packages/use/src/useRequest/demo/useRequest.vue
@@ -35,9 +35,7 @@
     isLoading,
     isFinished,
     isSuccess,
-    successCount,
     execute,
-    clearSuccessCount,
   } = useRequest(() => {
     return axios.get(url.value).then(res => omit(res, 'config'));
   }, {
@@ -56,12 +54,6 @@
     return (
       <pre class="scrollbar my-0 p-(y1 x3)">
         {yamlDump({ isExecuted, isLoading, isFinished, isSuccess })}
-        <div>
-          successCount:
-          {' '}
-          {successCount.value}
-          <button class="lh-none rounded el-3 ml-4 p-(x2 y1)" onClick={clearSuccessCount}>清除请求成功次数</button>
-        </div>
         {yamlDump({
           response: computed(() => response.value ?? `${response.value}`),
           data: computed(() => data.value ?? `${data.value}`),